server/admin-pw-service/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
pyjwt==2.10.1
